feat: implemented run container pipeline for dynamic container
